New API with getOrInsert and getOrInsertComputed
This updates the design to match what was presented to committee
at the October 2024 plenary. I am becoming champion of the proposal
with the students at the University of Bergen as co-authors, and the
previous champion and author are noted as no longer actively
involved in the proposal.

The naming of the methods is still up for discussion, I plan to
open an issue to discuss options once this pull request is merged.